refactor(ui): drop unused helpers, theme catalog, and chart barrels
