--- ui/src/composables/charts/index.ts
@@ -1,13 +1,0 @@
-export { useBarChartOptions } from './useBarChartOptions'
-export { useLineChartOptions } from './useLineChartOptions'
-export { usePieChartOptions } from './usePieChartOptions'
-export { useHeatmapChartOptions } from './useHeatmapChartOptions'
-export { useRadarChartOptions } from './useRadarChartOptions'
-export { useSankeyChartOptions } from './useSankeyChartOptions'
-export { useBar3DChartOptions } from './useBar3DChartOptions'
-export { useLine3DChartOptions } from './useLine3DChartOptions'
-export { useScatterChartOptions } from './useScatterChartOptions'
-export { useScatter3DChartOptions } from './useScatter3DChartOptions'
-export type { BaseChartConfig } from './baseChartOptions'
-export { getBaseOptions } from './baseChartOptions'
-export * from './shared'
