--- ui/src/composables/charts/shared/index.ts
@@ -1,9 +1,0 @@
-/**
- * Shared utilities for chart composables
- * Centralizes common functionality to reduce code duplication
- */
-
-export * from './chartConfig'
-export * from './seriesConfig'
-export * from './valueMode'
-export * from './3d'
