+++ ui/src/composables/charts/usePieChartOptions.ts
@@ -3,12 +3,8 @@ import type { EChartsOption } from 'echarts'
 import type { TitleComponentOption } from 'echarts/components'
 import { type BaseChartConfig, getBaseOptions } from './baseChartOptions'
 import { getNextColorFor, hasXAxis, hasYAxis, hasZAxis } from '@/lib/utils'
-import {
-  getChartStyling,
-  createPieSeriesConfig,
-  getTooltipTheme,
-  formatTooltipValue,
-} from './shared'
+import { getChartStyling, getTooltipTheme, formatTooltipValue } from './shared/chartConfig'
+import { createPieSeriesConfig } from './shared/seriesConfig'
 import { fontSize, sortByTotal, sortByValue } from './shared/common'
 import type { Point3D } from '@/types'
