+++ ui/src/composables/charts/useRadarChartOptions.ts
@@ -2,7 +2,7 @@ import { computed } from 'vue'
 import type { EChartsOption } from 'echarts'
 import { type BaseChartConfig, getBaseOptions } from './baseChartOptions'
 import { getNextColorFor, hasXAxis, hasYAxis, hasZAxis } from '@/lib/utils'
-import { getChartStyling, getTooltipTheme, formatRadarItemTooltip } from './shared'
+import { getChartStyling, getTooltipTheme, formatRadarItemTooltip } from './shared/chartConfig'
 import { fontSize, sortByTotal } from './shared/common'
 
 const makeIndicators = (names: string[], perSpokeMax: number[]) =>
