+++ ui/src/composables/charts/useScatter3DChartOptions.ts
@@ -2,6 +2,7 @@ import { computed } from 'vue'
 import type { EChartsOption } from 'echarts'
 import { type BaseChartConfig, getBaseOptions } from './baseChartOptions'
 import { getDefaultThemeColor, getNextColorFor } from '@/lib/utils'
+import { getChartStyling, getTooltipTheme } from './shared/chartConfig'
 import {
   EMPTY_RENDER,
   makeAxis3DCommon,
@@ -15,13 +16,11 @@ import {
   resolve3DVisualMap,
   createValue3DTooltipFormatter,
   buildContinuous3DOptions,
-  getChartStyling,
-  getTooltipTheme,
   makeContinuous3DParams,
   valuePoints3DToSeries,
   resolve3DZAxisType,
   type Continuous3DContext,
-} from './shared'
+} from './shared/3d'
 import { resolve3DSymbolProps } from './shared/seriesConfig'
 import { buildMixedAxes3DOptions } from './shared/mixedMode'
 import type { Series3DData } from '@/types'
